gpg: avoid deletion of keys not specified by user
When the user tries to delete a specific secret subkey
by using an exact search specification (trailing "!"),
gpg deletes the primary key and all other subkeys as well.

The new should_skip static function allows the key deletion routine
to determine when to skip keys. In this case, keys are skipped when
they are not the exact key specified by the user.

Signed-off-by: Matheus Afonso Martins Moreira <matheus.a.m.moreira@gmail.com>