gpg: confirm deletion of each secret key
* g10/delkey.c (confirm_deletion): New.
(do_delete_key): Confirm the deletion of each individual secret key.

--

When asked to delete a specific secret subkey, the program confirms
the deletion of the primary key once and then deletes the subkey.
It is confirming the deletion of the entire key block instead of
just the key the user specified.

Now it confirms the deletion of each individual secret key.

Signed-off-by: Matheus Afonso Martins Moreira <matheus.a.m.moreira@gmail.com>